[MAHOUT-1895] Move to MahoutConversions
